refactor: 느슨한 결합 구조로 변경 #7
